Refactor OCI artifact commands
