+++ cmd/kustomizer/pull_artifact_test.go
@@ -23,7 +23,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestPush(t *testing.T) {
+func TestPull(t *testing.T) {
 	g := NewWithT(t)
 	id := randStringRunes(5)
 	tag := "v1.0.0"
@@ -37,9 +37,20 @@ func TestPush(t *testing.T) {
 
 	t.Run("build and push artifact", func(t *testing.T) {
 		output, err := executeCommand(fmt.Sprintf(
-			"push %s -k %s",
+			"build -a %s -k %s",
 			artifact,
 			dir,
+		))
+
+		g.Expect(err).NotTo(HaveOccurred())
+		t.Logf("\n%s", output)
+		g.Expect(output).To(MatchRegexp(id))
+	})
+
+	t.Run("pull artifact", func(t *testing.T) {
+		output, err := executeCommand(fmt.Sprintf(
+			"pull artifact %s",
+			artifact,
 		))
 
 		g.Expect(err).NotTo(HaveOccurred())
