+++ cmd/kustomizer/tag.go
@@ -17,50 +17,14 @@ limitations under the License.
 package main
 
 import (
-	"context"
-	"fmt"
-
 	"github.com/spf13/cobra"
-
-	"github.com/stefanprodan/kustomizer/pkg/registry"
 )
 
 var tagCmd = &cobra.Command{
-	Use:   "tag OCIURL TAG",
-	Short: "Tag adds a tag for the specified OCI artifact.",
-	Long: `The tag command tags an existing image on the remote container registry.
-This command uses the credentials from '~/.docker/config.json'.`,
-	Example: `  # Tag an OCI artifact as latest
-  kustomizer tag oci://docker.io/user/repo:v1.0.0 latest
-`,
-	RunE: runTagCmd,
+	Use:   "tag",
+	Short: "Set artifacts tags to container registries.",
 }
 
 func init() {
 	rootCmd.AddCommand(tagCmd)
-}
-
-func runTagCmd(cmd *cobra.Command, args []string) error {
-	if len(args) != 2 {
-		return fmt.Errorf("you must specify an artifact name e.g. 'oci://docker.io/user/repo:tag' and a tag")
-	}
-
-	url, err := registry.ParseURL(args[0])
-	if err != nil {
-		return err
-	}
-
-	tag := args[1]
-
-	ctx, cancel := context.WithTimeout(context.Background(), rootArgs.timeout)
-	defer cancel()
-
-	res, err := registry.Tag(ctx, url, tag)
-	if err != nil {
-		return fmt.Errorf("tagging %s failed: %w", url, err)
-	}
-
-	logger.Println("tagged", res)
-
-	return nil
 }
